@@ -1,6 +1,6 @@
 module api-mapping-customization-guide
 
-go 1.19
+go 1.21
 
 require (
 	github.com/google/uuid v1.3.1
